requirements.txt
absl-py==0.3.0
astor==0.7.1
gast==0.2.0
grpcio==1.13.0
Markdown==2.6.11
numpy==1.15.0
opencv-python==3.4.2.17
protobuf==3.6.0
six==1.11.0
tensorboard==1.9.0
tensorflow-gpu==1.9.0
termcolor==1.1.0
Werkzeug==0.14.1
